refactor(react-ai-sdk): extract tool output conversion
